Added The Feature For Users to Pick a House + A Test
